Moved components and created new graphics
