+++ src/components/stopwatch-time/index.ts
@@ -1,2 +1,2 @@
-export * from "./stopwatch-time";
+export * from "../../graphics/scoreboard/src/components/stopwatch-time";
 export * from "./stopwatch-time-panel";
