+++ src/graphics/match.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <title>Handball Match</title>
+    <style>
+      body {
+        margin: 0;
+        padding: 0;
+      }
+
+      div#root {
+        width: 100vw;
+        height: 100vh;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+      }
+    </style>
+    <script type="module" src="./match/index.tsx"></script>
+  </head>
+  <body>
+    <noscript>This page needs javascript enabled</noscript>
+    <div id="root">Loading</div>
+  </body>
+</html>
