+++ src/graphics/match/app.tsx
@@ -1,0 +1,8 @@
+import { ReactElement } from "react";
+// import { SCOREBOARD_MAIN_TIMER } from "services/scoreboard-main-timer";
+
+function App(): ReactElement | null {
+  return <div></div>;
+}
+
+export default App;
