+++ src/graphics/scoreboard.html
@@ -3,17 +3,10 @@
   <head>
     <meta charset="utf-8" />
     <title>Handball Liveshow</title>
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Orbitron:wght@400;700&display=swap"
-      rel="stylesheet"
-    />
     <style>
       body {
         background-color: #00ff00;
         color: #151111;
-        /* font-family: "Orbitron", sans-serif; */
         font-family: "Monospace";
         font-size: 450px;
         font-weight: 700;
@@ -37,8 +30,18 @@
       div.scoreboard-time {
         font-size: 250px;
       }
+
+      div.score-separator {
+        font-size: 450px;
+        margin: 0 50px;
+      }
+
+      div.local-score-value,
+      div.visitor-score-value {
+        font-size: 450px;
+      }
     </style>
-    <script type="module" src="./layout-1/index.tsx"></script>
+    <script type="module" src="./scoreboard/index.tsx"></script>
   </head>
   <body>
     <noscript>This page needs javascript enabled</noscript>
