+++ src/graphics/scoreboard/src/components/scoreboard-results.tsx
