+++ src/graphics/scoreboard/src/components/stopwatch-time.tsx
